+++ ImcApp/ImcApp/App.xaml.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
+using ImcApp.Views;
 
 namespace ImcApp
 {
