--- ImcApp/ImcApp/MainPage.xaml
@@ -1,65 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<ContentPage xmlns="http://xamarin.com/schemas/2014/forms"
-             xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="ImcApp.MainPage">
-
-    <StackLayout>
-        <!--<Frame BackgroundColor="#2196F3" Padding="24" CornerRadius="0">
-            <Label Text="Welcome to Xamarin.Forms!" HorizontalTextAlignment="Center" TextColor="White" FontSize="36"/>
-        </Frame>
-        <Label Text="Start developing now" FontSize="Title" Padding="30,10,30,10"/>
-        <Label Text="Make changes to your XAML file and save to see your UI update in the running app with XAML Hot Reload. Give it a try!" FontSize="16" Padding="30,0,30,0"/>
-        <Label FontSize="16" Padding="30,24,30,0">
-            <Label.FormattedText>
-                <FormattedString>
-                    <FormattedString.Spans>
-                        <Span Text="Learn more at "/>
-                        <Span Text="https://aka.ms/xamarin-quickstart" FontAttributes="Bold"/>
-                    </FormattedString.Spans>
-                </FormattedString>
-            </Label.FormattedText>
-        </Label>-->
-        <Frame Margin="8" CornerRadius="0" BorderColor="LightGray">
-
-            <StackLayout>
-                <Label Text="Peso de la persona en Kilogramos:"
-                       FontSize="Body" />
-                <Entry FontSize="Medium" x:Name="PesoEntry"/>
-
-                <Label Text="Estatura de la persona en Metro:" 
-                       FontSize="Body"/>
-                <Entry FontSize="Medium" x:Name="EstaturaEntry"/>
-            </StackLayout>
-            
-        </Frame>
-
-        <StackLayout Orientation="Horizontal" Padding="4" 
-                     HorizontalOptions="CenterAndExpand">
-
-            <Button Text="Calcular" FontSize="Body" 
-                    WidthRequest="130" x:Name="CalcularEntry"
-                    Clicked="CalcularButton_Clicked"/>
-            <Button Text="Limpiar" FontSize="Body" 
-                    WidthRequest="130" x:Name="LimpiarEntry"
-                    Clicked="LimpiarButton_Clicked"/>
-
-        </StackLayout>
-
-        <Frame Margin="8" CornerRadius="0"
-               BorderColor="LightGray">
-
-            <StackLayout>
-                <Label Text="El indice de masa corporal de la persona es:" FontSize="Body"/>
-                <Label Text="23.4567" HorizontalOptions="Center"
-                       FontSize="28" x:Name="ImcLabel"/>
-                <Label Text="La situacion nutricional de la persona es:" FontSize="Body"/>
-                <Label Text="Peso Normal" HorizontalOptions="Center"
-                       FontSize="28" x:Name="SituacionNutricionalLabel"/>
-            </StackLayout>
-            
-        </Frame>
-        
-        
-    </StackLayout>
-
-</ContentPage>
